requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
psycopg2==2.7.4
pytz==2018.3
